test: Add test shell app.
This test application persists after launch and listens for commands on a socket; this enables a series of manipulations on a single container.

This should be useful for tests that need to perform an action in a container, then manipulate the broader flynn system, then verify the impacts from inside the container.

Signed-off-by: Eric Myhre <eric@flynn.io>