deployer: If the deployer fails, reschedule jobs.
The deployer might die at any point, leaving deploys in an incomplete state. To
remedy that, we track the state, and reschedule the jobs as needed. Note that
this does not currently rollback the deployment as our plan is to make
strategies idempotent, so the job can just continue where it left off. If we do
decide to rollback, we'll need to store the original formation in the deployment
itself, to be able to do the rollback.

Signed-off-by: Blaž Hrastnik <blaz.hrast@gmail.com>