test/rootfs: Disable memory overcommit
Setting this to ‘1’ effectively prevents errors from being returned from
malloc() and the OOM killer is never triggered. Thus, the CI VMs tend
to hang when they encounter memory leaks instead of causing issues that
would show up in the logs. Removing this customization reverts us back
to the Linux default of ‘0’ which behaves as expected.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>