controller: Upgrade StreamFormations handling.
`StreamFormations` now implements the pkg/stream pattern.  This should improve the ability to clean up after this function, and importantly, gives us a place to report errors should they arise in decoding (this fixes previous behavior, which was to simply drop such errors on the floor).

In the initial draft, I still had a 'formationStream' struct floating around, but after working through all the examples for the stream package and then coming back to this, it seems there's no unique details to necessitate such a structure (the only unique signature is in the data channel type itself, which of course the stream pattern excludes on purpose so as to avoid falling into a trap that requires generics).   So, I added stream.BasicStream instead, and just used that here.

Signed-off-by: Eric Myhre <eric@flynn.io>