discoverd2: Implement DNS server
This implements a DNS server backed by the discoverd database.
Queries that are not for the discoverd. domain will be proxied to
a recursive nameservers if configured.

Each registered service exposes a number of domains:

SERVICE

`SERVICE.discoverd.` returns randomly sorted A, AAAA, and SRV records
for each instance.

RFC 2782

`_SERVICE._PROTO.discoverd.` returns randomly sorted A, AAAA, and
SRV records for each instance, filtered by protocol.

LEADER

`leader.SERVICE.discoverd.` returns single A or AAAA and SRV records
pointing at the leader of the service.

INSTANCE

`INSTANCE_ID.SERVICE._i.discoverd` returns single A or AAAA and SRV
records for a single instance. This domain pattern is typically only
used as the target for SRV records (which include the port but not
the IP address).

Lookups over UDP are limited to up to three instances and the
response does not have the truncation bit set (which would trigger
a subsequent lookup over TCP). TCP lookups return all instances. The
weight and priority of SRV records is always 1. SRV lookups over TCP
include extra A/AAAA records matching the target.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>