router: make domain name lookup case insensitive
Some http clients do not downcase the hostname in the Host header of the
HTTP request and the servername in the TLS client-hello. Matching route
domains in a case-insensitive way prevents 404 and TLS handshake errors
for such clients. The request's Host header is not altered and keeps the
original casing.

Signed-off-by: Ben Burkert <ben@benburkert.com>