test: Add unit tests to flynn-test-runner.
Most of the work is done by 'test/scripts/test-unit.sh'.  flynn-test-runner also gained a short embedded wrapper script; this serves to set up the correct working directory, and give us a place to do autodetect for whether to run test-unit.sh (not all branches may have it).  A variety of alternatives to this embedded script were considered (including doing a loopback mount of the build vm filesystem on the host); this seems like the more direct route.  Fundamentally the reason this feels like an uphill battle is because we have code that we want to run per-branch, but some of it is embedded in a the flynn-test-runner which lives on a separate lifecycle; splitting these completely would remove that friction, but is not in reach for today.

If unit tests fail, the entire run will be halted; cluster bootstrapping and integration tests will not be run.

The unit tests are run within the build VM.  This is not by necessity, but seems like a reasonable time saving.

Signed-off-by: Eric Myhre <eric@flynn.io>