host: Expose nameservers to discoverd, use discoverd DNS
This shoehorns in the changes necessary to use discoverd DNS, by
parsing /etcd/resolv.conf from ConfigureNetworking, and then
injecting them into the manifest along with the bridge address.

This allows discoverd to bind to port 53 on the bridge address and
serve DNS to non-host networking container which get a resolv.conf
bind-mounted in that points at discoverd.

A side effect is that the IPv6 nameserver filter is removed, but
this is fine because only jobs with host networking see the host
resolv.conf.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>