router: Strip port from Host header when looking up route
It is totally valid for a Host header to look like:

    Host: example.com:80

or

    Host: example.com:443

This will cause a 404, as the route doesn't contain the port. So
when doing the backend lookup, we need to strip the port off if it
is included.

This fixes the "controller: resource not found" error that slipped
through broken CI.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>