router: Flush HTTP response headers and body writes
Without this, some of our streaming clients would hang forever
waiting for SSE events that would just hang out in the response body
buffer. It may make sense to switch to an interval instead of
flushing every write, but I want to get this in immediately, as it
is breaking a bunch of stuff.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>