router: proper handling of connection options
In RFC7235, the definition of hop-by-hop headers has changed somewhat.
Rather than always deleting a specific list of headers, the general rule
is to treat headers as hop-by-hop only if they were also specified as a
connection option in the Connection header. See Section 6.1 for more
info:

https://tools.ietf.org/html/rfc7230#section-6.1

Also, that same RFC extensively documents the Upgrade header field. In
particular, a proxy is supposed to remove that header before forwarding
the request if the Connection header doesn't also contain the "upgrade"
option, or if the client uses an HTTP version below 1.1:

https://tools.ietf.org/html/rfc7230#section-6.7

Finally, apply the same behavior to response headers.

Signed-off-by: Blake Gentry <blakesgentry@gmail.com>