test/cluster: Use virtio network interface
Refs #753

This resolves a long-standing networking instability issue on CI.
The default qemu NIC driver is apparently broken and was
dropping/mangling packets right and left. This patch changes the NIC
driver to virtio which is used by everybody and works correctly.

The problem and fix were extensively verified by running a program
that exercised the network by making many connections and
transferring many bytes.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>