router: properly determine fwd port from listener
Rather than just pulling the forwarding port out of the input address,
determine it based off of the listener's actual address. This makes sure
that the value isn't 0, even in tests where that was previously the
case.

Also use a mustPortFromAddr() helper to pull a port out of an address
string rather than using net.SplitHostPort everywhere.

Signed-off-by: Blake Gentry <blakesgentry@gmail.com>