host/volume: Volume manager state persistence.
Volume managers now use a boltdb (much like the host already uses to track jobs) to keep track of volumes across reboots of the flynn-host daemon.

Construction of a volume manager now takes a function that can be used to initialize a default volume provider, rather than just taking a provider reference; this makes it possible for the manager to only create a provider if it can't restore an existing provider already named "default".

This is a large set of changes relating to better separating data that should be persistent versus transient; other highlights include:

- `zfsVolume` now tracks its dataset.
- Mounts of volumes into containers are no longer tracked by individual volumes. (Since that functionality is the same for all volumes, there's no reason to duplicate it between implementations.)
- This also drove a change to make volumes publicly admit to their filesystem's mount location.  (Previously, we were attempting to conceal that detail, but as things have shaken out it seems increasingly clear that the final step of placing volume data in sight of a container is more the container backend's job.  Exposing this also makes testing much easier.)
- Another pleasant result of removing mount tracking from individual volumes is that cleanup from dead jobs is not the volume implementations' problem.
- Note that VolumeBinding instructions are *already* stored by the regular serialization of host.ActiveJob, so there is no delta visible there.  Currently, no special action is taken on that data on restore; presumably if the job survived a host reboot the volume mount did too.

Signed-off-by: Eric Myhre <eric@flynn.io>