host: Add cli flag for vol path.
This sets where the volume manager state db will be placed, and also the default locations for provider working dirs, which are also now configurable.

Renamed previous flag of same name for the older state system; this will be removed in the near future when it has been fully supplanted.

Signed-off-by: Eric Myhre <eric@flynn.io>