host/volume/zfs: Rename test file.
This was originally contained more general stuff, but at this point snapshot tests have their own suite, and this rename is a more accurate description of the file's contents.

Additionally, as long as we're getting large diffs out of the way, refactor to remove now-unnecessarily zfs-specific casts and calls from the tests, and rename the import to 'gzfs' as most other files already do to reduce the potential for ambiguity.

Signed-off-by: Eric Myhre <eric@flynn.io>