host/volume: Snapshot permanence.
Snapshots are now a type of volume that is retained (previously, we just forked new writable volumes again using snapshots as a temporary intermediate, thus requiring no special treatment of snapshots).  They should behave almost entirely the same as other volumes, but we both want to keep them, and also need to retain the distinction because zfs treats them differently (namely, in send/recv, which are upcoming).

This means snapshots are now read-only (tests have been updated to match) and getting a writable volume is a two step process using `CreateSnapshot` and `ForkVolume`.

Any operation involving mutations to volumes (i.e. creating snapshots) is being moved to functions on the provider.  This simplifies bookkeeping; volumes are more a data access object.

Signed-off-by: Eric Myhre <eric@flynn.io>