appliance/postgresql: Implement cluster state machine
This commit adds a Go translation of the manatee-state-machine
project[0] from Joyent. It is translated verbatim with the following
minor differences:

- A few extra tests were added for completeness
- Changes were to translate from the Node.js async model to a more
  straightforward model that uses channels.
- The term 'one-node-write-mode' (abbreviated as ONWM) was replaced
  with 'singleton' as they are used interchangeably upstream.

The state machine provides a fault-tolerant system for running
Postgres clusters that automatically fail over with minimal downtime
and no possibility for data loss or split brain during failovers.

A simulator is used for testing and is also available in an
interactive REPL for experimenting with various scenarios.

[0] https://github.com/joyent/manatee-state-machine/tree/d441fe941faddb51d6e6237d792dd4d7fae64cc6

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>