bootstrap: Run highly available clusters when not in singleton mode
This changes bootstrap to detect whether it should run in singleton mode
by first checking the SINGLETON environment variable and if not set,
checking if the number of expected nodes is 1.

If singleton mode is not enabled, enough instances of each process are
run to make the control plane tolerate the loss of a node without having
to start new instances.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>