cli: Add psql command for access to postgres consoles
This adds a new `flynn psql` command that opens a console to the Flynn
Postgres database configured for the application.

It retrieves the postgres details from the app’s current release and
then uses the release from the referenced postgres service for a one-off
job that runs the psql command.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>