controller: Retry requests after dial errors
There is a chance that service discovery could return an address of a
controller and that instance subsequently goes down, meaning when a
HTTP request is actually made, there will be an i/o timeout.

Previously the client would use a connect timeout of 30s (set by
http.DefaultTransport) which is too long to wait in this case.

This change lowers the connect timeout to 1s, and retries dialling the
controller multiple times if an error occurs, giving up after 30s.

Signed-off-by: Lewis Marshall <lewis@lmars.net>