discoverd: Fix manual leader when restarting discoverd
If a manual leader is in the backing datastore and the instance is
not registered when we start, we should notify and store the leader
if/when that instance registers.

This fixes an issue where restarting discoverd resulted in the
manual leader not being used.

Also fixed are some internal state consistency issues when removing
instances with manual leader election.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>