host/containerinit: Don’t use pkg/shutdown
Since we need to trap all signals and pass them along to our child, we
can’t use the shutdown package as it waits for signals as well and will
terminate the process. We don’t use the shutdown hooks and stderr isn’t
hooked up to anything, so we can just replace the usage with os.Exit.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>