host: Stop when stopping libvirt-bin
flynn-host depends on libvirt-bin, so we should shutdown cleanly
before it stops.

This requires a line to be added to /etc/default/libvirt-bin to
ensure that the pre-stop script in /etc/init/libvirt-bin does not
stop containers out from under us:

  libvirt_uris='qemu:///system'

That line overrides a variable in the libvirt-bin init script that
lists lxc as well and is used to list+shutdown in pre-stop.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>