deployer: Don't wait for 0 job events
waitForJobEvents compares the expected jobEvents with the actual events
received, and does not treat `map[typ:map[down:0]]` the same as actual
not having a `typ` key, so instead don't set 0 so that expected will be
"equal" to actual.

This presented itself when a release had 0 running processes of a given
type (because they crashed), so the all-at-once deployment would end up
waiting for 0 down events. Ideally we would have a test for this, but
due to the scheduler having a restart policy for crashed jobs, it would
make a test not 100% reliable (there may be no crashed jobs as the
scheduler just restarted them).

Signed-off-by: Lewis Marshall <lewis@lmars.net>