discoverd: Wait for flannel interface to come up
discoverd binds to port 53 on the flannel interface in order to
provide DNS to containers. When resurrecting, discoverd may start
before flannel is finished setting up the interface.

This patch uses a startup script that parses the expected IP from
the arguments and waits for it to show up in ipconfig before
exec'ing discoverd.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>