host,pkg/cluster: Add SignalJob
Signalling a job can be achieved via the attach protocol but only when
also requesting logs. Sending a signal without requesting logs currently
fails as the remote end hangs up the connection, but in this case
establishing an attach connection is overkill, so I've added a simpler
host API method.

Signed-off-by: Lewis Marshall <lewis@lmars.net>