all: Build our own cross-platform Go toolchain
This ensures that we build with a consistent version of Go and that we
can cross-compile the CLI for all supported platforms.

Use the Makefile to build it with gonative instead of tup, because tup
can’t handle not knowing all of the output files and it’s not reasonable
to list the entire output.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>