test: Increase deploy timeout to 60s in TestDeployController
The deployer can sometimes take a while to reconnect to the controller
after deploying it. We should investigate why it takes so long, but for
now bumping the timeout to give it more chance to succeed, as if the
deploy eventually succeeds after the test fails, other tests break as
the controller is still being deployed.

See the following CI build log for an example of the timeout not being
long enough:

https://s3.amazonaws.com/flynn-ci-logs/20150312210925-4961845f-build-8aed616b0248792c51ce448a807370da982560da-2015-03-12-21-18-03.txt

Signed-off-by: Lewis Marshall <lewis@lmars.net>