cli: Update using TUF
This removes the old distribution toolchain inherited from hk and
replaces it with an updater that uses TUF. The same TUF repo that is
used to distribute the rest of the container images and binaries is
used.

script/export-components is updated to add the cross-compiled flynn
CLI binaries to the repo tagged with their version. The CLI will be
released and versioned as part of the nightly build process instead
of using a separate version tag and build system.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>