slugrunner: Fix command quoting
This fixes running commands with simple quoting and spaces like:

    flynn run bash -c "echo 'foo bar'"

I have not tested more advanced cases, as I suspect that this is
good enough for the majority of use cases (even this does much
quoting does not work on Heroku).

The test is in the git deploy suite because it requires an SSH key
to be set up and we don't use git anywhere else in the test suite.

Closes #656

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>