installer: Fix UI glitches (windows)
Setting `-moz-appearance: none` used to be broken but has been fixed, and
the work-around of setting it to radio-container is now broken on
windows.

The hack for removing the dotted focus border on `select`s in firefox
breaks in chrome on windows, so only apply it for firefox.

Added Meiryo to list of fonts so the `-` renders correctly in chrome on
windows (integer picker).

fixes #1353

Signed-off-by: Jesse Stuart <jesse@jessestuart.ca>