deployer: Start new Postgres instances before stopping old ones
The deployer needs Postgres to be read-write (or eventually read-write)
so that it can modify formations and get the scheduler to start jobs,
but stopping an Async before starting a new one will leave the Postgres
cluster read-only as there always needs to be an Async for read-write.

By starting new instances before killing old ones, this situation is
avoided as there is always an Async present.

Signed-off-by: Lewis Marshall <lewis@lmars.net>