all: Add resource specifications
This implements what was proposed in #1496 except for hosts declaring
available resources.

You will notice calls to resource.SetDefaults in the controller / host
code which I am using to explicitly fill in the resource specifications
(rather than blank ones meaning default) so that it is clear to the end
user from job / release config what the limits are, defaults or not.

Signed-off-by: Lewis Marshall <lewis@lmars.net>