controller: Use a single DB connection for app events
Previously, each event stream request would open a DB connection, and it
seems some of these streams leak in CI which leads to the following
error:

```
pq: remaining connection slots are reserved for non-replication
superuser connections
```

This commit moves to a single DB connection for all app events, and
individual streams subscribe and receive events from a single in-process
listener loop, which doesn't fix the leak, but at least does not lead to
leaked DB connections.

Each subscriber gets an in-memory queue of events so that slow receivers
don't block the listener loop.

Signed-off-by: Lewis Marshall <lewis@lmars.net>