host: Remove default max procs limit
The nprocs rlimit limits the process count of the user id of the process
it is applied to. This makes it useless currently, as most jobs end up
running as either root or nobody, resulting in errors like:

    fork: Resource temporarily unavailable

When there are more than nproc process running for a specific uid. In
practice, this can happen when scaling a single app to a dozen or so
instances.

The long-term solution is to make sure every job has a different uid,
but this is a complex undertaking, so for now we’ll just remove the
default limit.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>