host: Clean up mounts that leak into libvirt_lxc
When a mount namespace is created, it inherits all of the mounts
from its parent. If they are not umount(2)'ed, they will remain and
prevent devices from being cleaned up. This currently causes
problems when destroying ZFS volumes and cleaning up temporary
container rootfs mountpoints.

The libvirt-lxc container system creates two mount namespaces for
each container: one in the libvirt_lxc supervisor process, and one
in the process that execs the actual container process. The latter
cleans up the mounts that leak in, but the former does not. This is
tracked upstream at
https://bugzilla.redhat.com/show_bug.cgi?id=1067489

We work around this issue by adding a small flynn-nsumount
executable that can reach into an arbitrary mount namespace (using
a pid) and call umount(2) against a list of paths. This executable
is called immediately after launching each libvirt-lxc container to
clean up any leaked mountpoints that are known to cause trouble with
flynn.

Note that this is not a complete fix, just a temporary band-aid. The
following issues remain:

- Any devices such as USB drives will not be automatically cleaned
  up, as we only unmount paths that we know we control. This could
  probably be solved, but the hope is that this workaround will not
  be required for very long.
- There is a non-zero time period between creating the namespace and
  unmounting the leaked mounts, which could cause a race with
  processes that are destroying mounts. In practice this probably
  won't happen and can be worked around by retrying requests.

Closes #1125

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>