gitreceived: Fix directory permissions issues
The change in e8652e235d81eed79ded998a24ad1a34b607fe84 to checkout a
git working directory instead of using git-archive caused a permissions
handling bug that mangled permissions and omitted the +x flag from some
directories when the working directory was restored from the repo cache.

This was fixed by stripping the (redundant) checked out working directory
when creating the cache. A test was added to verify that the permissions
bug does not return.

Closes #1757

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>