gitreceived,cli: Run git over HTTPS instead of SSH
This changes gitreceived to use the Smart HTTP git protocol instead
of SSH, which removes the need for users to manage SSH keys.

The CLI now configures two global git options that pin the CA
certificate for the git remote and use a credential helper built
into the CLI to get the password. In the future this helper will be
used to support per-user auth, etc.

Port 2222 is no longer used and is removed from all configs/docs.

The CLI cluster config is simplified, but remains temporarily
backwards-compatible with SSH gitreceive, we'll remove the backwards
compatibility when few users have old versions deployed.

Closes #1744

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>