scheduler: Bugfixes and refactors for syncing (host, job, and formation)
scheduler: No need to handle jobEvents closing

It is an internal channel which never gets closed.

Signed-off-by: Lewis Marshall <lewis@lmars.net>

scheduler: Only log events if there are listeners

Signed-off-by: Lewis Marshall <lewis@lmars.net>

scheduler: Refactor followHost

Signed-off-by: Lewis Marshall <lewis@lmars.net>

scheduler: Ensure we are in sync before starting the loop

Starting the loop before being fully in sync with the hosts and
formations leads to a lot of churn on startup.

We also need to only register in service discovery once we are in sync
with the hosts so that a deployment doesn't kill old schedulers before
the new schedulers are watching events, which could lead to events being
missed.

Signed-off-by: Lewis Marshall <lewis@lmars.net>

scheduler: Rename IsEqual -> Equals

Signed-off-by: Lewis Marshall <lewis@lmars.net>

scheduler: Don't track one-off jobs

Signed-off-by: Lewis Marshall <lewis@lmars.net>

scheduler: Update logger in HandleFormationChange

Signed-off-by: Lewis Marshall <lewis@lmars.net>