scheduler: Ensure we are in sync before starting the loop
Starting the loop before being fully in sync with the hosts and
formations leads to a lot of churn on startup.

We also need to only register in service discovery once we are in sync
with the hosts so that a deployment doesn't kill old schedulers before
the new schedulers are watching events, which could lead to events being
missed.

Signed-off-by: Lewis Marshall <lewis@lmars.net>