cli,host/cli: Implement full-cluster backup and restore
`flynn cluster backup` saves a full-cluster backup to a local tar file.
The backup consists of expanded app details for the discoverd, flannel,
postgres, and controller apps and a SQL dump of all databases in the
postgres cluster.

`flynn-host bootstrap --from-backup` bootstraps a cluster and restores
a backup. First the discoverd, flannel, and postgres apps are started on
the cluster, then the SQL is restored along with some extra commands
that update the artifact URLs to the latest ones from the bootstrap
manifest.

After the data is restored, the controller/scheduler is started and the
scheduler handles starting the rest of the Flynn and user apps.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>