installer: Use a separate EBS volume for data
This avoids the initialization cost of EBS volumes provisioned from
snapshots, which all boot volumes are. The initial read(2)/write(2)
of a section of an EBS volume can take many seconds due to an underlying
read from the snapshot in S3.

This was the cause of discoverd instance expiry, heartbeat writes were
taking many seconds while the volume was initializing.

The solution also has the added benefit of not using a sparse file
layered on top of ext4 for the ZFS pool.

Closes #2027
Closes #2060

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>