appliance/postgresql: Add persistent instance ID
This patch adds a persistent instance ID that is saved into the postgres
data volume, which allows the volume to be used for cluster restoration.

The instance ID is matched against the service metadata in discoverd
that includes details about instance roles. Using an ID linked to the
data directory allows the proper role to be created. Starting and instance
with the data directory from a Primary or Sync will “do the right thing.”

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>