controller: Add App.DeployTimeout
Both the controller and postgres can take longer than the default 30s to
deploy, and so we have some custom timeouts spread amongst the code, but
the default timeout is used in cases like `flynn -a controller env set`,
so is susceptible to timeouts.

I have added App.DeployTimeout which is the timeout in seconds, and
added Deployment.DeployTimeout which is copied from the app so the
client knows what the timeout is when creating a deployment without also
looking up the app.

Signed-off-by: Lewis Marshall <lewis@lmars.net>