controller: Fix hang in controller tests
The tests were haging during TestStreamDeployment, and I noticed that
pgx was blocking trying to acquire a connection. Looking at the code, we
were creating transactions before reading from the database, and so I
moved the transactions and it no longer hangs.

I don't know the specific reason this was an issue (I assume reading
inside a transaction is an ok thing to do?).

Signed-off-by: Lewis Marshall <lewis@lmars.net>