controller: Avoid double close panic during shutdown
During shutdown, c.eventListener.CloseWithError(ErrShutdown) is called
this causes the pg listener channel to be closed which will race to be
handled and could result in another call to CloseWithError which causes
another close of e.doneCh.

To avoid this we check if the listener is closed in CloseWithError
and skip the close logic if it has already been closed.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>