test: Increase the deploy timeout
The scheduler occasionally does not receive job events from hosts, so
only updates the job status in the controller when a sync happens.
Bumping the deploy timeout to be greater than the sync period ensures
deployments don't timeout because of this.

We should probably investigate why some events are being dropped, see
for example DeployerSuite.TestOmniProcess in the following build:

https://s3.amazonaws.com/flynn-ci-logs/20160207043827-c2d85abb-build-1542c912ced512b896132a9b99eaa51ffcff1fde-2016-02-07-05-00-21.txt

Signed-off-by: Lewis Marshall <lewis@lmars.net>