pkg/httpclient: Increase stream reconnection timeout to 20s
Most streams are to the controller, and sometimes the controller can
misbehave during deployments and not shut down correctly (e.g. waiting
for DB connections to finish), but is eventually killed after 10s.

Also having a stream reconnect timeout of 10s means there is a chance we
are timing out connecting to a failing controller which is just about to
be replaced.

Signed-off-by: Lewis Marshall <lewis@lmars.net>