host: Add CPU limit option for jobs
CPU limits are implemented as a resource denominated in “milliCPUs”, a
concept from Google. 1000 milliCPUs is equivalent to a single CPU core.

A limit of 1000 milliCPUs results in the Linux default of 1024 CPU shares,
and the request portion of the resource spec can be used for scheduling
in the future.

CPU shares are relative, the more shares a process has, the higher
priority it is. When a host is under load, a job with 2000 milliCPUs
will get twice the CPU time as a  job with the default of 1000.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>