host: Add resource partitions with CPU quotas
Resource partitions complement CPU limits in order to isolate system
apps from user apps. A top-level set of cgroups is added that all
jobs are nested under. Currently these cgroups just provide CPU
share allocations, but in the future they can be used for any other
cgroup control. The partitions are only throttled when all
partitions are under equal load, full bursting is allowed, unused
shares are given to the other partitions on a proportional basis.
With that in mind, the default allocations are:

- `system`, all Flynn-managed system apps, with 25% of available CPU time
- `user`, all user app processes, with 50% of available CPU time
- `background`, slugbuilder builds, with 25% of available CPU time

Allocations can be modified by passing the `--partitions` flag to
`flynn-host daemon` like this:

    flynn-host daemon --partitions "system=cpu_shares:4096 background=cpu_shares:4096 user=cpu_shares:8192"

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>