scheduler: Include restarted jobs when counting host jobs
job.restartTimer is non-nil for any job that was a restart of another
job, even when that job is running. This means that checking for a nil
job.restartTimer causes GetHostJobCounts to ignore jobs which are
restarts of others, regardless of whether those jobs are running or not,
and thus can lead to too many jobs running on a single host.

This caused errors in a user's cluster because omni jobs were being
placed unevenly, leading to the surplusOmniJobs routine stopping them,
but then anti-entropy then placing the job unevenly again, leading to
the omni job flapping indefinitely.

Fixes #2757.

Signed-off-by: Lewis Marshall <lewis@lmars.net>