discoverd: Don't error if DNS isn't running during promote/demote
If the DNS server hasn't been started yet it's likely we are
are waiting on the host network to be configured.
Because we hold the lock after waitHostDNSConfig() unblocks the DNS
configuration will block on us releasing the lock.
Once this happens the DNS server will be properly configured taking into
account the state after promotion/demotion.

Signed-off-by: Joseph Glanville <jpg@jpg.id.au>