scheduler: Don't rely on formation / job event order
When a formation change happens, jobs are started and stopped in
goroutines which means there is no guarantee the tests will get
formation events before job events.

In most cases, the test just cares that a job event occurs. The fact
that it is due to a formation change is irrelevant, so there is no need
to explicitly wait for the formation event.

Signed-off-by: Lewis Marshall <lewis@lmars.net>