scheduler: Make job event order explicit in tests
TestMultipleHosts starts and stops jobs whilst the scheduler is paused
and then relies on the scheduler seeing the started job before the
stopped one. This is so that the number of jobs involved is predictable
(if the scheduler sees the stopped job first, it starts a replacement
job, then sees another job started and kills the one it just started).

Since the scheduler has a goroutine per host, there is no guarantee
which goroutine will handle it's event first, so a hook has been added
to ensure the event from AddJob has made it to the scheduler before
returning.

Signed-off-by: Lewis Marshall <lewis@lmars.net>