all: Enable TCP keepalive for all TCP listeners
By default on Linux, TCP connections in the ESTABLISHED state do not
get reaped for 5 days:

    $ sysctl net.netfilter.nf_conntrack_tcp_timeout_established
    net.netfilter.nf_conntrack_tcp_timeout_established = 43200

This means that if a FIN or RST packet is not received, for example
if a user closes their laptop instead of shutting down their browser
cleanly, the server will keep blocking on the fd for a very long
time, resulting in what appears to be a slow fd leak.

The solution is to enable TCP keepalives on the socket which will
tell the kernel ping the client periodically and if no response is
received, will move the connection into the CLOSE_WAIT state,
freeing the fd and server resources that were serving it.

The creator of the socket has to opt-in to this. By default net/http
does this when listening, but there are a variety of places where we
don't rely on net/http to set up the socket. This patch wraps the
listener in each of those cases so that keepalive is enabled.

Closes #2684

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>