host: Load network/discoverd configuration after crash
If flynn-host terminates without stopping jobs, when it reconnects
it needs to load the discoverd and network configuration back in so
that new jobs can be started.

Before this patch, newly added jobs would block forever waiting for
networking to come up.

The details are persisted in the global backend state storage
system, and a set of functions are set up to ensure it is kept up to
date when the relevant job IDs change and so on.

This also resulted in a cleanup of state restoration on updates, the
same code path is now used.

To test, I did the following:

1. Sent a SIGKILL to flynn-host and then started it back up, and
   confirmed that new jobs could be scheduled and the appropriate
   log messages were printed.
2. Sent a POST /host/update to simulate an in-place upgrade and
   confirmed that new jobs could be scheduled and the appropriate
   log messages were printed.
3. Stopped flynn-host cleanly and then started it back up and
   confirmed that the old config was not used, the appropriate log
   messages were printed, and the cluster resurrected.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>