host: Fix resurrection deadlock
During resurrection, the state database is open for reading, so
SetStatusFailed (which tries to update the state database) will blocks
until the resurrection has completed, but then that causes a deadlock
with UnmarshalState which is waiting to receive the error on the ready
channel, so this change sends the error before making the blocking
SetStatusFailed call.

Fixes #3232.

Signed-off-by: Lewis Marshall <lewis@lmars.net>