host: Fix checking minimum update version
Updating from versions prior to v20160711.0 is currently broken as they
try and download the nsumount binary which no longer exists.

We actually decided not to support in-place updates from libvirt-lxc to
libcontainer so we just need to tell users to do a backup/restore
instead, so I have bumped the minVersion and also fixed it so that it's
compared against the version of hosts running in the cluster (rather
than checking its own version, which will never fail).

Signed-off-by: Lewis Marshall <lewis@lmars.net>